use CarbonView from nef fw
