+++ nef/AppDelegate.swift
@@ -62,7 +62,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     }
     
     private func preferencesDidFinishLaunching() {
-        window = NSWindow(contentRect: NSRect(x: 0, y: 0, width: 800, height: 760),
+        window = NSWindow(contentRect: NSRect(x: 0, y: 0, width: 800, height: 768),
                           styleMask: [.titled, .closable, .miniaturizable],
                           backing: .buffered, defer: false)
