+++ nef/Assembler.swift
@@ -23,7 +23,7 @@ class Assembler {
     func resolveOpenPanel() -> OpenPanel { OpenPanel() }
     
     func resolveCarbon(code: String, output: URL) -> IO<AppDelegate.Error, URL> {
-        let model = Carbon(code: code, style: preferencesDataSource.state.style)
+        let model = Carbon(code: code, style: preferencesDataSource.state.carbonStyle)
         return nef.Render.build.carbonIO(model, output: output).mapLeft { _ in .carbon }
     }
