+++ nef/Extensions/CarbonStyle+Model.swift
@@ -4,7 +4,7 @@ import Foundation
 import nef
 
 extension PreferencesModel {
-    var style: CarbonStyle {
+    var carbonStyle: CarbonStyle {
         CarbonStyle(background: color,
                     theme: theme,
                     size: size,
